+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/ClientRetryPolicyE2ETests.java
@@ -14,6 +14,7 @@ import com.azure.cosmos.CosmosDiagnosticsContext;
 import com.azure.cosmos.CosmosException;
 import com.azure.cosmos.DirectConnectionConfig;
 import com.azure.cosmos.FlakyTestRetryAnalyzer;
+import com.azure.cosmos.TestObject;
 import com.azure.cosmos.implementation.AsyncDocumentClient;
 import com.azure.cosmos.implementation.DatabaseAccount;
 import com.azure.cosmos.implementation.DatabaseAccountLocation;
@@ -21,7 +22,6 @@ import com.azure.cosmos.implementation.GlobalEndpointManager;
 import com.azure.cosmos.implementation.HttpConstants;
 import com.azure.cosmos.implementation.OperationType;
 import com.azure.cosmos.implementation.Utils;
-import com.azure.cosmos.implementation.throughputControl.TestItem;
 import com.azure.cosmos.models.CosmosBatch;
 import com.azure.cosmos.models.CosmosChangeFeedRequestOptions;
 import com.azure.cosmos.models.CosmosItemIdentity;
@@ -196,7 +196,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
 
     @Test(groups = { "multi-master" }, dataProvider = "preferredRegionsConfigProvider", timeOut = TIMEOUT)
     public void queryPlanHttpTimeoutWillNotMarkRegionUnavailable(boolean shouldUsePreferredRegionsOnClient) {
-        TestItem newItem = TestItem.createNewItem();
+        TestObject newItem = TestObject.create();
 
         CosmosAsyncContainer resultantCosmosAsyncContainer;
         CosmosAsyncClient resultantCosmosAsyncClient;
@@ -239,7 +239,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
         try {
             // validate the query plan will be retried in a different region and the final requests will be succeeded
             // TODO: Also capture all retries for metadata requests in the diagnostics
-            FeedResponse<TestItem> firstPage = cosmosAsyncContainerFromClientWithPreferredRegions.queryItems(query, queryRequestOptions, TestItem.class)
+            FeedResponse<TestObject> firstPage = cosmosAsyncContainerFromClientWithPreferredRegions.queryItems(query, queryRequestOptions, TestObject.class)
                 .byPage()
                 .blockFirst();
 
@@ -274,7 +274,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
             throw new SkipException("queryPlanHttpTimeoutWillNotMarkRegionUnavailable() is only meant for DIRECT mode");
         }
 
-        TestItem newItem = TestItem.createNewItem();
+        TestObject newItem = TestObject.create();
         resultantCosmosAsyncContainer.createItem(newItem).block();
 
         // create fault injection rules for address refresh
@@ -309,8 +309,8 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 resultantCosmosAsyncContainer,
                 Arrays.asList(addressRefreshDelayRule, serverGoneRule)).block();
         try {
-            CosmosItemResponse<TestItem> itemResponse = resultantCosmosAsyncContainer
-                .readItem(newItem.getId(), new PartitionKey(newItem.getId()), TestItem.class)
+            CosmosItemResponse<TestObject> itemResponse = resultantCosmosAsyncContainer
+                .readItem(newItem.getId(), new PartitionKey(newItem.getId()), TestObject.class)
                 .block();
 
             assertThat(itemResponse).isNotNull();
@@ -379,7 +379,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 resultantCosmosAsyncContainer,
                 Arrays.asList(addressRefreshDelayRule, serverGoneRule)).block();
         try {
-            TestItem newItem = TestItem.createNewItem();
+            TestObject newItem = TestObject.create();
             resultantCosmosAsyncContainer.createItem(newItem).block();
         } catch (CosmosException e) {
             assertThat(e.getDiagnostics().getContactedRegionNames().size()).isEqualTo(1);
@@ -417,7 +417,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
             throw new SkipException("queryPlanHttpTimeoutWillNotMarkRegionUnavailable() is only meant for GATEWAY mode");
         }
 
-        TestItem newItem = TestItem.createNewItem();
+        TestObject newItem = TestObject.create();
         resultantCosmosAsyncContainer.createItem(newItem).block();
         FaultInjectionRule requestHttpTimeoutRule = new FaultInjectionRuleBuilder("requestHttpTimeoutRule" + UUID.randomUUID())
             .condition(
@@ -506,7 +506,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
             throw new SkipException("leaseNotFound is only meant for Direct mode");
         }
 
-        TestItem createdItem = TestItem.createNewItem();
+        TestObject createdItem = TestObject.create();
 
         FaultInjectionRule leaseNotFoundFaultRule = new FaultInjectionRuleBuilder("leaseNotFound-" + UUID.randomUUID())
             .condition(
@@ -619,7 +619,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
         CosmosFaultInjectionHelper.configureFaultInjectionRules(testContainer, Arrays.asList(channelAcquisitionExceptionRule)).block();
 
         try {
-            TestItem createdItem = TestItem.createNewItem();
+            TestObject createdItem = TestObject.create();
             testContainer.createItem(createdItem).block();
 
             // using a higher concurrency to force channelAcquisitionException to happen
@@ -683,21 +683,21 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
     private Mono<CosmosDiagnostics> performDocumentOperation(
         CosmosAsyncContainer cosmosAsyncContainer,
         OperationType operationType,
-        TestItem createdItem,
-        Function<TestItem, PartitionKey> extractPartitionKeyFunc,
+        TestObject createdItem,
+        Function<TestObject, PartitionKey> extractPartitionKeyFunc,
         boolean isReadMany) {
 
         try {
             if (operationType == OperationType.Query && isReadMany) {
                 CosmosQueryRequestOptions queryRequestOptions = new CosmosQueryRequestOptions();
                 String query = String.format("SELECT * from c where c.id = '%s'", createdItem.getId());
-                FeedResponse<TestItem> itemFeedResponse =
-                    cosmosAsyncContainer.queryItems(query, queryRequestOptions, TestItem.class).byPage().blockFirst();
+                FeedResponse<TestObject> itemFeedResponse =
+                    cosmosAsyncContainer.queryItems(query, queryRequestOptions, TestObject.class).byPage().blockFirst();
                 return Mono.just(itemFeedResponse.getCosmosDiagnostics())
                     .onErrorResume(throwable -> {
                         if (throwable instanceof CosmosException) {
                             CosmosException cosmosException = (CosmosException) throwable;
-
+                            
                             return Mono.just(cosmosException.getDiagnostics());
                         }
                         return Mono.error(throwable);
@@ -717,7 +717,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                         .readItem(
                             createdItem.getId(),
                             extractPartitionKeyFunc.apply(createdItem),
-                            TestItem.class
+                            TestObject.class
                         )
                         .map(itemResponse -> itemResponse.getDiagnostics())
                         .onErrorResume(throwable -> {
@@ -759,7 +759,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 }
 
                 if (operationType == OperationType.Create) {
-                    return cosmosAsyncContainer.createItem(TestItem.createNewItem()).map(itemResponse -> itemResponse.getDiagnostics())                        .onErrorResume(throwable -> {
+                    return cosmosAsyncContainer.createItem(TestObject.create()).map(itemResponse -> itemResponse.getDiagnostics())                        .onErrorResume(throwable -> {
                         if (throwable instanceof CosmosException) {
                             CosmosException cosmosException = (CosmosException) throwable;
 
@@ -770,7 +770,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 }
 
                 if (operationType == OperationType.Upsert) {
-                    return cosmosAsyncContainer.upsertItem(TestItem.createNewItem()).map(itemResponse -> itemResponse.getDiagnostics())                        .onErrorResume(throwable -> {
+                    return cosmosAsyncContainer.upsertItem(TestObject.create()).map(itemResponse -> itemResponse.getDiagnostics())                        .onErrorResume(throwable -> {
                         if (throwable instanceof CosmosException) {
                             CosmosException cosmosException = (CosmosException) throwable;
 
@@ -788,7 +788,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                             createdItem.getId(),
                             extractPartitionKeyFunc.apply(createdItem),
                             patchOperations,
-                            TestItem.class)
+                            TestObject.class)
                         .map(itemResponse -> itemResponse.getDiagnostics())
                         .onErrorResume(throwable -> {
                             if (throwable instanceof CosmosException) {
@@ -823,8 +823,8 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 CosmosChangeFeedRequestOptions changeFeedRequestOptions =
                     CosmosChangeFeedRequestOptions.createForProcessingFromBeginning(feedRanges.get(0));
 
-            FeedResponse<TestItem> firstPage =  cosmosAsyncContainer
-                .queryChangeFeed(changeFeedRequestOptions, TestItem.class)
+            FeedResponse<TestObject> firstPage =  cosmosAsyncContainer
+                .queryChangeFeed(changeFeedRequestOptions, TestObject.class)
                 .byPage()
                 .blockFirst();
             return Mono.just(firstPage.getCosmosDiagnostics());
@@ -834,7 +834,7 @@ public class ClientRetryPolicyE2ETests extends TestSuiteBase {
                 return cosmosAsyncContainer
                     .readMany(
                         Arrays.asList(new CosmosItemIdentity(extractPartitionKeyFunc.apply(createdItem), createdItem.getId()), new CosmosItemIdentity(extractPartitionKeyFunc.apply(createdItem), createdItem.getId())),
-                        TestItem.class)
+                        TestObject.class)
                     .map(itemResponse -> itemResponse.getCosmosDiagnostics())
                     .onErrorResume(throwable -> {
                         if (throwable instanceof CosmosException) {