+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdRequestHeaders.java
@@ -132,6 +132,7 @@ final class RntbdRequestHeaders extends RntbdTokenStream<RntbdRequestHeader> {
         this.addChangeFeedWireFormatVersion(headers);
         this.addPriorityLevel(headers);
         this.addGlobalDatabaseAccountName(headers);
+        this.addThroughputBucket(headers);
 
         // Normal headers (Strings, Ints, Longs, etc.)
 
@@ -290,6 +291,8 @@ final class RntbdRequestHeaders extends RntbdTokenStream<RntbdRequestHeader> {
     }
 
     private RntbdToken getPriorityLevel() { return this.get(RntbdRequestHeader.PriorityLevel); }
+
+    private RntbdToken getThroughputBucket() { return this.get(RntbdRequestHeader.ThroughputBucket); }
 
     private RntbdToken getGlobalDatabaseAccountName() {
         return this.get(RntbdRequestHeader.GlobalDatabaseAccountName);
@@ -789,6 +792,16 @@ final class RntbdRequestHeaders extends RntbdTokenStream<RntbdRequestHeader> {
                     .getPriorityLevelAccessor()
                     .getPriorityValue(priorityLevel)
             );
+        }
+    }
+
+    private void addThroughputBucket(final Map<String, String> headers)
+    {
+        final String value = headers.get(HttpHeaders.THROUGHPUT_BUCKET);
+
+        if (StringUtils.isNotEmpty(value)) {
+            final int throughputBucket = Integer.valueOf(value);
+            this.getThroughputBucket().setValue((byte)throughputBucket);
         }
     }